+++ .github/workflows/e2e-tests.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Set up Go
         uses: actions/setup-go@v6
         with:
-          go-version: "1.26"
+          go-version: "1.26.1"
 
       - name: Set up Node.js
         uses: actions/setup-node@v4
