+++ docs/benchmarking/run-your-own-benchmarks.mdx
@@ -23,7 +23,7 @@ Want to see Bifrost's performance in your specific environment? The [**Bifrost B
 
 Before running benchmarks, ensure you have:
 
-- **Go 1.23+** installed on your testing machine
+- **Go 1.26.1+** installed on your testing machine
 - **Bifrost instance** running and accessible
 - **Target API providers** configured (OpenAI, Anthropic, etc.)
 - **Network access** between benchmark tool and Bifrost
