+++ examples/mcps/http-no-ping-server/go.mod
@@ -1,6 +1,6 @@
 module http-no-ping-server
 
-go 1.26
+go 1.26.1
 
 require github.com/mark3labs/mcp-go v0.43.2
