+++ examples/plugins/http-transport-only/go.mod
@@ -1,6 +1,6 @@
 module github.com/maximhq/bifrost/examples/plugins/http-transport-only
 
-go 1.26
+go 1.26.1
 
 replace github.com/maximhq/bifrost/core => ../../../core
