+++ framework/go.mod
@@ -1,6 +1,6 @@
 module github.com/maximhq/bifrost/framework
 
-go 1.26
+go 1.26.1
 
 require (
 	github.com/google/uuid v1.6.0
