+++ plugins/governance/go.mod
@@ -1,6 +1,6 @@
 module github.com/maximhq/bifrost/plugins/governance
 
-go 1.26
+go 1.26.1
 
 require gorm.io/gorm v1.31.1
