+++ plugins/semanticcache/go.mod
@@ -1,6 +1,6 @@
 module github.com/maximhq/bifrost/plugins/semanticcache
 
-go 1.26
+go 1.26.1
 
 require (
 	github.com/cespare/xxhash/v2 v2.3.0
