+++ tests/governance/go.mod
@@ -1,3 +1,3 @@
 module github.com/maximhq/bifrost/tests/governance
 
-go 1.26
+go 1.26.1
