+++ tests/scripts/migration-checker/go.mod
@@ -1,3 +1,3 @@
 module github.com/maximhq/bifrost/tests/scripts/migration-checker
 
-go 1.26
+go 1.26.1
